Remove unnecessary utils package initializer
